Support updating for macOS builds.
We previously only supported Linux builds. We now have a PR out for
building the macOS distribution as well. This means that we need to also
support updates for macOS.

This adds a compile time check for the various operating systems we
support to ensure we are pulling the correct binary.