Fix update issue for tmpfs
There was an issue moving update files after they had been downloaded
and the signatures had been verified on tmpfs. Update files would be
downloaded and temporarily stored in `/tmp` which existed on tmpfs. Any
attempts to *move* the files to the disk (a separate file system) would
trigger a panic due to crossing boundaries.

This update changes to a *copy* and remove operation from the temporary file
location.

Also added an additional signature verification that occurs after the
files have been moved to the final install location. This is to address
a possible race condition where an attacker may change the file in
`tmp/` after the signature verification has occurred, but before we have
copied the file to the final location.