Use proper name during update
Previously, when an update occurred we were moving the binary to the
final install location after verifying the signatures. We changed this
to a move in a recent change to ensure that we could avoid crossing
filesystem boundaries in the event a user was using tmpfs or similar.

The name in the copy ended up being the binary asset name, which took on
the form of `phylum-<os>-<arch>`. However. the installer (and the user)
expects this to be `phylum`. This change simply ensures that we use the
`phylum` name for binaries after update/install.