Add email to `phylum auth status`
Previously the `auth status` command would only check that the token is
present in the configuration file, however it would neither confirm its
validity nor print information about the user of that token.

By requesting the user's information from the authentication API, it's
possible to both provide more insight about the user currently logged in
and verify that the token is actually valid instead of just present.

Currently only the user's E-Mail is displayed, since this is the
information used for logging in.

Closes #174.