Add uninstall script
This adds a new script to the published archives which allows for easy
uninstallation of the phylum CLI.

This implementation intentionally deviates from the behavior suggested
in #206, since removal of configuration files is usually not expected by
the user. So only files generated by the installer will be removed.

To ensure it's still possible to clean up all phylum CLI files including
configurations, the `--purge` option is available.

Closes #206.