Add Python poetry.lock support
This adds support for analyzing the Python `poetry.lock` package lock
format.

Since this format includes a metadata field with a version indicator, a
warning is emitted when a breaking change that isn't supported by
`phylum` was detected.

While it would be possible to abort completely, it's not unlikely that a
format change does not touch the information required by `phylum` at
all. So we attempt to continue anyway and just let the user know that
some packages might not have been analyzed.

Closes #237.