Add uninstall subcommand to phylum
In #236 a shell script was proposed to allow for uninstallation of the
Phylum CLI, however the usage of that command turned out to be a bit
awkward.

This patch instead introduces the new `phylum uninstall` CLI option
which will automatically remove the CLI without having to go through the
hassle of acquiring a separate uninstallation script.

To ensure uninstallation never removes something it shouldn't, the shell
configs are only edited when the full text added by the installer can be
found. To ensure this works as often as possible, the `install.sh`
script had to be modified to always include the completion
initialization, since this is commonly present already.

Closes #206.