Add yarn v2 lockfile support
This adds support for the new yaml-based yarn lockfile introduced with
yarn version 2.

Similar to the new package-lock.json, the yarn lockfile now annotates
the different packages with where they were resolved from. Currently
only packages that resolve to npm are submitted for analysis.

To detect which version of the yarn lock is present, the entire file is
first attempted to parse with `serde_yaml` and if that fails, the v1
parser is used as fallback instead.

Closes #246.