Replace Language/Type with Ecosystem
Since a package uploaded to NPM might not necessarily be JavaScript, the
existing `phylum analyze` description could be misleading.

To improve on this, the `Language` and `Type` fields have been
consolidated into an `Ecosystem` field which just prints the package
repository rather than a guess at the language used in the package.

Closes #129.