Add support for new npm package-lock format (#242)
This adds initial support for the new package-lock.json format
introduced in version 7, where a new `packages` section was introduced
with the goal of replacing `dependencies`.

One major change with the new `dependencies` section is that a full path
to each dependency is specified, rather than just the dependencies name
(e.g. "node_modules/accepts" vs. "accepts"). To accomodate for this,
everything before the last `node_modules/` in the package identifier is
truncated, returning only the trailing part of the path.

While extracting the module name out of the path has worked in all
examples I've tested, the ideal solution would likely be to rely on the
`resolved` field to acquire the dependencies directly rather than
relying on name resolution.

Closes #153.