Truncate excessive project names
Since a lot of output in the Phylum CLI relies on tables, it's important
that text does not exceed the maximum cell length. The most likely
culprit for this is the `project`, which can be freely defined by the
user.

All instances of the project name output are now truncated to the
respective cell length and a truncation is indicated using the ellipsis
symbol.

To ensure this works for project names in any language, the length of
the cells is calculated using the `unicode-width` crate.

Closes #127.