Set appropriate exit codes on failure (#260)
Several paths of the Phylum CLI would print warnings/errors to the user,
but still set and exit code of `0`. This makes it difficult for for
automated tools to verify success and can confuse users.

As a solution, the `CommandValue` enum has been changed, replacing the
`CommandValue::Void` variant with `CommandValue::Code`. This code is
then used on exit instead of the fixed `0` which was used for
`CommandValue::Void`.

This also introduces a few simplifications to the error/CommandValue
handling, removing unnecessary code and inlining unnecessary
abstractions. This hopefully improves transparency of the process and
makes the usage consistent.

Some functions relying on `CommandValue::Void` as a replacement to `()`
without ever propagating it have been switched to `anyhow::Result<()>`.
Generally there's little reason to use `CommandValue` unless the goal
is for it to end up directly as the command's exit code.

Some functions also had to be adjusted to ensure the code is properly
propagated. This should capture all instances where the result of a
function returning `CommandResult` was ignored.

Closes #229.