Simplify the PR template
This patch removes the boilerplate from our pull request template. What
is left is only the checklist, which is important to ensure that nothing
is missed before a PR is merged.

If developers are using detailed commit messages (like this one), PRs
will automatically be created with the detailed commit message followed
by the checklist, which is exactly how we want them.