chore: refresh the release script
While releasing the last handful of CLI versions it became clear that
the release script used to do so could use a refresh. For one, it
automated _everything_, which felt risky and most of the time the script
would be aborted before it could tag and push the changes. Another
minor annoyance was not remembering the format of the version or
changelog entry. Additionally, the git log summary would only show from
the last tag, not taking into account that the last tag could be from a
pre-release.

These issues have been addressed, along with some minor
refactoring and formatting. It is still just a basic shell script and
there could certainly be better ways yet.