Fix yarn lockfile parsing with patched deps (#343)
Yarn allows patching dependencies, which produces a local patch file
applied on top of a dependency. Before this patch, this would cause an
error in the lockfile parser.

To provide the most value possible, this patch adds support for parsing
these patched dependencies and extracts the original package, while
discarding the local patch information. This allows analyzing the
unpatched dependency for vulnerabilities.

Closes #342.