Rewrite yarn v2 parser
This is a complete change in how the yarn v2 parser parses the
resolution field, which should hopefully both make the parser simpler
and more error-resistent.

Instead of complex splitting logic, the parser now makes use of the fact
that dependencies inclued at most one `@` at the beginning of their
name, thus clearly identifying the resolver as anything past the second
`@`.

The supported resolvers are now explicitly handled by checking the
format of the resolution's resolver, an unrecognized resolver will cause
an error.