Remove uuid dependency
It turns out that we don't actually need this as a direct dependency
because `phylum_types` has type aliases for all of this.

Doing it this way has pros and cons. Maybe someday we'll decide that the
aliases in `phylum_types` don't actually make sense or that they should
actually use the newtype pattern. But for now, they exist, so we might
as well use them.

As an nice bonus, this means we won't have to worry about coordinating
our version bump of the uuid crate with phylum-dev/phylum-types#27. Once
that is merged, we will automatically be upgrade (after the next cargo
update)