Fix yarn v1 parser with quoted version key
This resolves an issue with the parser for the old v1 yarn format where
a lockfile generated by npm instead of yarn would cause all keys in the
lockfile to be quoted, breaking the parser.

This patch updates the parser to accept optional quotes around the
`version` key in the lockfile, since this is the only optionally quoted
field we care about.

Closes #380.