Add support for `gradle.lockfile`
Before this patch, only preliminary Gradle support was present in the
Phylum CLI by using some of Gradle's output as pseudo-lockfile. This
patch adds proper Gradle support by adding a parser for the
gradle.lockfile file format.

Since the previous format was unreliable and based on human-readable
output, the new Gradle parser replaces the old one instead of being
offered as an alternative.

Closes #372.