Add deno runtime for executing extensions
This adds a deno runtime which is capable of executing the `entry_point`
of an extension as an ES module.

This supports both traditional JavaScript modules and transpiled
TypeScript without any additional dependencies for the user.

Module imports are resolved through the filesystem relative to the entry
point, allowing for multiple source files to be shipped with a single
executable.

Closes #417.