Use rust-cache for test workflow
* Use rust-cache for test.yml

Update test.yml to use rust-cache in order to speed up builds.

* Don't run the release workflow for pull requests

The release workflow is not a required check for PRs and it creates
artifacts that waste space without providing much benefit. As we speed
up the test workflow, it is more likely that PRs will be merged before
the release workflow is done. And the release workflow can, of course,
be manually run if there is ever any question.

* Use v3 of actions/checkout instead of master