Fix stack overflow on Windows
This resolves an issue when trying to run the CLI on Windows, which
would cause a stack overflow due to the clap CLI parsing.

To reduce the amount of stack necessary to parse the CLI, all instances
of the `arg!` macro have been replaced with manually built `clap::Arg`
arguments.

Closes #424.