Unify XDG directory methods
This merges together the two separate `data_dir` methods and streamlines
all the other XDG-related utility functions.

Since the only error that is possible with these methods is the very
uncommon scenario where the user's home directory does not exist, error
handling of this scenario has been made less explicit and prominent.

Due to the current order of operations, the errors should still be the
same. However this would make it more likely that in the future an
uninstall is only partially attempted due to a missing home directory.

Closes #414.