feat: coordinate phylum-ci Docker image releases with new CLI releases
Add a repository dispatch event trigger to the release workflow. It
triggers the Docker workflow in the `phylum-dev/phylum-ci` repo to build
and push Docker images containing the latest `phylum` Python package and
the Phylum CLI version from the tag that caused the Release workflow to
run. The trigger is a simple curl command that POSTs to the `dispatches`
endpoint with the expected payload. The trigger is further limited so
that it will not fire for pre-releases of the CLI.