Combine duplicate ignore_certs fields (#479)
This patch unifies the two instances of `ignore_certs` in the
`PhylumApi`. One of which was stored on `Config`, while the other was
stored directly on `PhylumApi`.

Now `Config::ignore_certs` is the unambiguous truth for determining cert
considerations and has potential CLI overrides applied directly after
deserialization.

This also removes the unused `PhylumApi::packages` field.