fix: phylum-ci Docker image trigger
The change in this PR is to encapsulate the `curl` based trigger's data
payload in double quotes since expressions do not expand in single
quotes. This means that the inner double quotes have to be escaped
because they are still needed since the payload includes a JSON string.

Promote the `if` condition for triggers from step to job level.
This change will ensure the Trigger job does not run for pre-releases.
Previously, the conditional was one level lower, on a step within the
job. That was confusing because the UI would show that the Job ran, but
it wasn't until drilling into the results that it would be clear that
the trigger step was skipped.

Closes # 488