Add extensions integration tests
This adds a set of extension integration tests, which run against the
staging environment.

Since these tests can take some time, they're hidden behind the
`integration-tests` feature flag, which can also be used for future
long-running integration tests.

To ensure these don't break, they're still run on every commit in CI,
but if execution time is excessive that can be changed in the future
with only changes to the CI.

Closes #429.