Make job ID mandatory in job status extension api
Before this patch, the `get_job_status` extension API would fall back to
the project ID when no job ID was specified, however the project ID is
not a valid job ID.

Since there's no reasonable default for us to pick, this patch instead
switches the job ID parameter to be mandatory.

This also removes the alias for "current" as the latest job ID, since
this API changed due to the request manager removal.