Prevent extension install with invalid permissions
Since Deno's check for networking permissions only considers the domain,
the scheme and path can be disregarded and should be stripped to prevent
confusing users when prompting for permissions.

While it would be possible to strip this and continue while warning the
user, preventing an install ensures that no extensions with invalid net
permissions are ever published. It also avoids printing warnings when
deserializing existing extensions during startup for clap's CLI.

Closes #519.