Move extension API behind `Rc`
By wrapping the extension API behind an `Rc`, with the only way to
access it through the `ExtensionState::from` method, it should be
impossible to hold a `Ref` to `OpState` across any await points.

The API of `ExtensionState` should provide a nice safe wrapper to
consumers which never have to worry about the intricacies of holding
state across await points.

The internal `ExtensionStateInner` and its `state` module provides a
dedicated place for potentially hazardous operations and comes with a
detailed warning helping developers steer clear of potential pitfalls.