Refactor CLI output formatting (#564)
This patch combines the two `Summarize` and `Renderable` traits into a
single `Format` trait, to simplify the CLI outputting process.

Instead of filtering issues as part of the `Summarize` trait, there's
now also a separate `FilterIssues` trait which allows filtering issues
on demand.

The `print_response` method responsible for handling errors and
outputing JSON/summarize based on pretty flag was removed in favor of
anyhow error propagation in combination with `Format::write_stdout`.

The `--filter` flag was also added to the `phylum package` subcommand,
since this was the only subcommand that outputs issues which didn't
allow for issue filtering.

Additionally this fixes a bug in the histogram where the bar for buckets
with only a single package would not show up at all. This could mislead
developers into thinking there are no vulnerabilities in a project. Now
there will always be at least a one cell wide bar.

Closes #106.