Fix running extensions with relative paths
This patch fixes an issue with `phylum extension run`, where relative
paths were not accepted as parameter.

To ensure that our module loader's restriction to only load from within
the extension's directory still works properly, all paths are not
canonicalized in the module loader. This means we do not need to check
for symlinks explicitly.

Closes #586.