Add git dependency support to package-lock.json
This resolves an issue with the v7+ package-lock.json parser where
remote git dependencies would be submitted as npm packages, causing the
analysis to never complete.

This patch uses the `resolved` field instead of `version` for all
packages that do not resolve to the `registry.npmjs.org` domain.

This leaves the v6 lockfile parser unaffected, since backporting
shouldn't be necessary given that most people should move on to v7+
automatically over time.

Closes #622.