Fix PHYLUM_API_KEY overwriting config token (#631)
This fixes an issue with the CLI where the config's token would be
overwritten whenever using the PHYLUM_API_KEY environment variable.

Instead of overwriting the config based on what is stored in the
environment variable, the variable and config value are now tracked
separately, always writing the config's value whenever a write is
instructed and returning the environment variable (if present) when
reading from it.

Closes #582.