Add -musl shims to release archives
This adds glibc releases as *-unknown-linux-musl to the release archives
to allow older versions to seamlessly transition to the new glibc
executables.

This does NOT provide a musl version, the archives are still glibc
executables and this will only help with updates on systems that support
both musl AND glibc.