Move error logging to API module
Instead of handling the addition of higher-level detail to error
messages in the locations where the API is used, this moves it to the
API instead. This way all locations calling these API functions will
implicitly get the refined error messages without requiring additional
handling at a higher level.

This change was inspired by the existing `get_project_id` function,
which already provides a friendlier error message for missing projects.
So this patch also helps with consistency of error handling