Fix parsing of nested package-lock dependencies
This fixes an issue with npm package-lock.json dependencies where
packages placed in other dependencies' node_modules folders were not
correctly identified.

Instead of parsing `node_modules/dep/node_modules/dep2` as
`dep/node_modules/dep2`, it is now identified properly as `dep2`.

This kind of package structure can occur when dependencies exist
multiple times in the dependency graph. Placing it as a child of a
specific dependency is npm's way to keep multiple package versions,
since the directory on the root does not have the version as part of its
name. This is why both versions must be submitted.

Closes #662.