Add CLI flags for log level control
This adds the new `-v` and `-q` CLI flags which allow controlling the
log level for the `env_logger`.

The default log level is `warn`, logging both errors and warnings. The
`-q` flag can be specified up to two times to disable logging completely
while the `-v` flag can be specified up to three times to enable
trace-level logging.

Closes #722.