Add `phylum init` subcommand
This adds a new subcommand for setting up a new Phylum project.

Currently this command will only negotiate the desired project name and
corresponding group, but further information like lockfile can be
detected in the future.

Project and group can either be specified directly in the CLI, or
specified interactively.

The user is not prompted for a group name when passing the project name
on the CLI, to enable non-interactive usage without passing extra CLI
flags. So in this scenario a group must be specified with `--group`.

Closes #795.