Add lockfile path and type to .phylum_project
This patch extends the existing .phylum_project with two new optional
fields: `lockfile` and `lockfile_type`.

When the `lockfile` type is present in the .phylum_project file and no
lockfile override is specified on the CLI, then the project file from
the .phylum_project file will be used with the lockfile type specified
in the same file.

If the lockfile type is specified on the CLI without the lockfile being
specified, it will be ignored.

The lockfile name and type can be set when running `phylum init`, either
by passing `--lockfile` and (optionally) `--lockfile-type`, or by
specifying them interactively.

Closes #793.