Add group member management subcommands
This patch adds the new `phylum group members` subcommand which has the
additional `list`/`add`/`remove` subcommands. These allow managing and
inspecting the users which are part of a group.

The `members`/`list` commands behave similarly to all our other
commands, doing the same thing and just listing all members that are
part of the specified group.

The group parameter is a required parameter for the `members`
subcommand, which allows clap to provide helpful messages guiding the
user to specify it in the correct location, rather than having to handle
all possible locations internally.

Closes #TBD.