Fill out copyright notice in license templates
While not necessary, this fills out the copyright information in the
license. This could help point out Phylum's involvement should the
license ever get packaged with another application.

The main argument for adding this additional information however is
consistency with phylum-dev/birdcage. This way all of our licenses have
the same structure.