Add automatic CLI doc generation
This patch adds a new `clap_markdown` crate which automatically converts
our internal `clap::Command` constructed for the CLI into a markdown
page compatible with reame.io's markdown renderer.

Generation is done using `cargo run --bin xtask -- gendocs` or with the
`cargo xtask gendocs` alias.

A test was added which validates in CI that the currently generated
documentation is always up to date with the latest CLI, reminding
developers to update the documentation when making any changes to
Phylum's CLI.

Closes #803.