Ignore parent directory projects for `phylum init`
Previously, `phylum init` would query if the project should be
overwritten when a parent project contained a project file already.

Since we do not actually overwrite any project files outside of the
current directory, and since having nested Phylum projects is perfectly
valid, a warning is now only emitted when there's already a Phylum
project file in the current directory.

Closes #839.