Fix subdir analysis without lockfile parameter
Since the lockfile was specified as a relative path in the project
lockfile, an analysis would fail when the analysis was started from a
child directory of the Phylum project directory.

To ensure that analysis will work in any of the child directories, a
root path of the project is now stored in the config on load and the
lockfile location is determined based on that root directory whenever it
is required.

While an absolute path to the lockfile might work temporarily, it was
intentionally avoided as a solution since it would break when the
project is moved to a different location or shared between different
systems.

Closes #843.