Reorder project initialization
This reorders the steps in the project initialization for the `phylum
init` command, to improve user experience with the command.

Instead of trying to create or link to the project before prompting for
the lockfile path and type, all API operations are now moved to the end
after the user has finished his inputs. This way looking can go on in
the background without users waiting for long-running API calls to
complete.

Closes #841.