Remove unused zip compression algorithms
Since we have full control over the compression algorithm used to create
the Phylum executable, it is not necessary to support multiple
algorithms.

The only necessary algorithm is deflate, which is what our zip files are
compressed with by default.

This should improve compile times, since zstd-sys compilation in CI
takes over 2 minutes.