Restore files on ecosystem extension API failure
This patch resolves an issue with the ecosystem extensions where the dry
run could potentially leave the project manifest and lockfile modified
if the extension failed a `PhylumApi` call, since this introduces an
exception.

To ensure backups are always restored after the dry run, we catch the
exception, restore and then throw it again. Throwing the same exception
has the advantage that error formatting is persisted.

This also fixes an issue in all three ecosystem extensions where calls
to the async `abort` function were not awaited. As a result it was
possible that execution continued despite explicitly trying to abort.

Closes #860.