Fix group prompt during `phylum init`
Currently when running `phylum init`, the user is first asked if they
want to use a group and then forced to enter a group name. While it has
documented none sa the default, it is not possible to just press enter
to use that default. As a result users are forced to abort and restart
`phylum init` if they accidentally landed here.

This patch combines the prompt for entering a group and the actual group
name input into a single prompt. This way the user is never "stuck" in a
state where he has to Ctrl+C out of. This also makes it quicker to enter
a group should it be desired.