Fix parser lockfile consistency
This patch is a rewrite of the lockfile parser with the goal to improve
the consistency between the individual parsers.

Previously, some of the lockfile parsers would just ignore things like
version ranges in favor of trying to extract as much valid data as
possible. Instead any unknown syntax is now an immediate parsing failure
and only strict dependency versions are allowed. This way we only ever
analyze proper lockfiles.

Another inconsistency was the treatment of different dependency types.
Some parsers were submitting git dependencies while others were just
ignoring them. To help with parsing these different version types,
there's a new `PackageVersion` enum in place of a version string which
encodes where the version came from.

Using the new `PackageVersion` enum, it is now possible to filter
packages in the CLI and determine after parsing which packages should be
submitted. This also allows for printing warnings in the future to
inform users about potentially unanalyzed dependencies.