Fix sandbox executable path resolution
Previously we were resolving relative paths through `$PATH` by adding
their path to each entry in `$PATH` and checking if the resulting path
exists. This could easily lead to wrong results for paths like `.` which
would always exist when adding it to any path in `$PATH`.

To ensure path resolution only works when the full path specified is
actually an executable in any directory in `$PATH`, we walk the `$PATH`
directories instead and check if its entries' file names matches the
requested path exactly.