Fix gem parser for dependencies without version
The ruby parser expected a loose version for all dependencies and
sub-dependencies. However it is not required for sub-dependencies to
have any version at all.

To fix this, the loose version parser has been extended to allow for a
completely empty string as an alternative to a version and the tests
were modified to ensure this is tested for.

Closes #918.