Add option to specify multiple lockfiles on CLI
This patch changes the `--lockfile` option in all places to allow for
multiple occurrences, allowing proper multi-lockfile support for the
entire CLI.

The `--lockfile-type` parameter itself is still only allowed once and
applies to all other specified lockfiles. This should be fine
considering it's only a last-ditch escape hatch since it should always
be possible to just let the auto-parser figure things out.

Closes #906.