Fix poetry extension
This patch updates the poetry extension to make use of the `--lock` flag
for dry-running the package installation without executing any code.

Installation after analysis is done outside of the sandbox, since it
seems to allow for near arbitrary code execution.

A more blacklist-based approach to deny some access (like $HOME) could
likely be introduced in the future.

Closes #804.