Fix install with poetry extension
Since the `poetry install` subcommand does not have a `--lock` argument,
it would cause the subcommand to fail. Since `install` should generally
be based on an updated lockfile already anyway, it should be fine to run
an analysis without first updating the lockfile.

However there are options allowing `install` to modify the lockfile,
which is why a proper solution would likely involve parsing the
`--dry-run` output instead.