Suggest available groups for `phylum init`
Previously the group input field was just a freetext field, which
created issues when the group did not exist already since groups are not
automatically created.

To resolve this, we send a request for available groups during startup
now and only allow selecting groups the user is part of.

Without any groups, the group selection process will be skipped
entirely.

The groups HTTP request is sent on startup since we have some startup
delay anyway and this way it will not harm the interactivity after this
initial startup period.

Closes #934.