Fix Go parser ignoring dependencies
Previously the assumption was made that any line in the Go lockfile
including `/go.mod` after its version was only included for the hash and
would still have its version included without the `/go.mod` in a
different line. This assumption was incorrect.

To ensure we don't ignore dependencies that are actually used, we now
include all versions with and without the suffix and deduplicate that
list after parsing.

This will include a lot more dependencies than what is actually used,
but for security purposes over-reporting is much likely more desirable
than under-reporting, considering we can't get an accurate list just by
parsing the lockfile.

Closes #942.