ci: improve the release process (#952)
The `softprops/action-gh-release` third party action was removed. It was replaced with a GitHub-maintained action, `actions/github-script`, to create the release and a direct use of the API via `curl` to upload the release assets.

Additionally, the use of the signing key was modified so that it is no longer written to disk.

Finally, some light refactoring was done, mostly to adhere to `shellcheck` findings, as discovered by running the excellent `actionlint` tool.

All of these changes were tested in a separate repository first...in order to actually exercise the workflow with releases without affecting the `phylum-dev/cli` repository. To see these changes in action...with the history of trial and error...check out the [maxrake/delme](https://github.com/maxrake/delme) repository. It is currently public but may be converted to private and/or deleted soon.

Closes #951