Use --dry-run output for poetry extension
This fixes some issues with the poetry ecosystem extension by switching
from updating the lockfile and parsing it over to parsing the
`--dry-run` flag's output instead.

The primary issue with the `--lock` flag is that the `poetry install`
subcommand does not have such a flag, only `poetry add`.

Besides `poetry install`, this patch also removes any necessity to write
to the project directory since `--dry-run` is immutable. This allows
further lockdown of our sandbox.

Closes #931.