Switch default subcommand from `list` to `help`
Currently commands like `phylum project` default to `phylum project
list` when no subcommand is specified. This means that all flags present
in `phylum project list` also need to be available on `phylum project`,
which makes it confusing to know where the user is supposed to specify
them.

To reduce the potential for confusion, the command `phylum project` is
now just an alias for `phylum project help`. Having fewer ways to do the
same thing will hopefully make it more clear what the correct approach
is without having to rely on extensive documentation and error messages.

While it would be possible to just print an error listing the available
subcommands, this patch chooses to print the help instead since it's a
less aggressive and more helpful output, especially considering the user
likely made this mistake because he's not aware of how the command
should be used in general.

Since the `phylum auth` command was relying on a custom implementation
for essentially the same functionality, it has been migrated to the
built-in clap flag too.

Closes #812.
Closes #555.