Add pip extension as official extension (#980)
* Add pip extension from community-extensions repository

* Resolve type issues in pip/main.ts

* deno fmt pip/main.ts

* Inherit stderr during `pip install --dry-run`

Using "piped" stderr leads to confusing error messages. For example:

```
> phylum pip install some-pkg-that-doesnt-exist
[phylum] Finding new dependencies…
[phylum] Pip dry-run failed.
```

Allowing stderr to pass through is better:

```
> phylum pip install some-pkg-that-doesnt-exist
[phylum] Finding new dependencies…
ERROR: Could not find a version that satisfies the requirement some-pkg-that-doesnt-exist (from versions: none)
ERROR: No matching distribution found for some-pkg-that-doesnt-exist
[phylum] Pip dry-run failed.
```

* Mention correct repository in README

* Note pip version requirement

* Add pip extension to official extensions

* deno fmt pip/README.md

Because apparently deno fmt cares...

* Use pip install --dry-run --report

Parsing the dry run output is unnecessary. Pip provides a machine
readable [installation report].

[installation report]: https://pip.pypa.io/en/stable/cli/pip_install/#obtaining-information-about-what-was-installed

* Remove --ignore-installed flag from pip install --dry-run

I misunderstood this flag when I added it to the command. We do NOT want
to ignore the installed dependencies. On the contrary, by considering
the dependencies that are already installed, we avoid checking packages
that aren't in the diff.

* deno fmt pip/main.ts

* Add pip extension to extension overview

* Fix yarn link in extension docs

* Note version requirement for pip install --report

* Sort issues by severity

* Confirm the version of pip's install report

* Fix repot typo

* Mention correct pip version requirement in pip extension README