Switch to policy endpoint for job results
This patch switches to the new policy endpoint for `phylum analyze` and
`phylum history <JOB_ID>`.

Since the thresholds-based analysis endpoint is not functional anymore,
only the new endpoint is used and exposed to the extension API. This
makes it a breaking change for all extensions using the `getJobStatus`
function.

Closes #1002.